requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.50.0
google-genai>=1.20.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
numpy>=1.24.0